Basic typecasting is performed.
You can once again pass an int or a float to a function expecting a
float or a double. Arrays are not cast (I don't think they ever were: I
tested to see if adding a derive for them would work, and it didn't).

I also eliminated all the warnings given by *warn-on-reflection*.