Allows cloggle to be compiled without breaking.
Previously, if cloggle had been compiled, using it would cause a fatal
error. This was caused by making flush private and putting gl-flush in
its place, and the way compilation initialises Vars. See:
http://groups.google.com/group/clojure/browse_thread/thread/74d7f7216fc9355b
This works around the issue with a hack which is not as ugly as it might
be.